% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicagoad.R
\name{chicagoad}
\alias{chicagoad}
\title{Generate a PDF Document with the Chicago AD Template}
\usage{
chicagoad(...)
}
\arguments{
\item{...}{Arguments passable to \code{rmarkdown::word_document}}
}
\value{
A PDF document with the Chicago AD template
}
\description{
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it was originally intended for use with Word, not PDF, which has much more limited functionality for templates. Please use \code{draft_chicago} instead.
}
\details{
Thank you to Steve Miller's \code{stevetemplates} library for providing a baseline for this function.
}
\examples{
try(rmarkdown::render("path/to/your.Rmd", kevins.templates::chicagoad()))
}
\keyword{internal}
