% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_chicago.R
\name{draft_chicago}
\alias{draft_chicago}
\title{Draft a new R Markdown document using the Chicago AD style}
\usage{
draft_chicago(name, ...)
}
\arguments{
\item{name}{String value for the name of the new R Markdown document (i.e., "my_new_document.Rmd")}

\item{...}{Additional arguments to pass to \code{rmarkdown::draft}}
}
\value{
A new R Markdown document using the Chicago AD style
}
\description{
Draft a new R Markdown document using the Chicago AD style
}
\examples{
try(kevins.templates::draft_chicago("my_new_document.Rmd"))
}
