% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevins.templates-package.R
\docType{package}
\name{kevins.templates-package}
\alias{kevins.templates}
\alias{kevins.templates-package}
\title{kevins.templates: Provides RMD Templates for Class Essays}
\description{
This package will serve as a home for any RMD templates I'll ever need or create.
}
\author{
\strong{Maintainer}: Kevin Navarrete-Parra \email{navark4@unlv.nevada.edu}

}
\keyword{internal}
