% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word}
\alias{word}
\alias{templ_word}
\title{Kevin's Chicago/Apsa style template}
\usage{
word(...)

templ_word()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:word_document]{rmarkdown::word_document}}.}
}
\description{
This is an R markdown template that makes it easier to generate word documents that
follow the Chicago/APSA format rules. Accordingly, the template uses the author-date
version of Chicago format.
}
