% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicagoad.R
\name{chicagoad}
\alias{chicagoad}
\title{Generate a PDF Document with the Chicago AD Template}
\usage{
chicagoad(...)
}
\arguments{
\item{...}{Arguments passable to \code{rmarkdown::word_document}}
}
\value{
A PDF document with the Chicago AD template
}
\description{
Generate a PDF Document with the Chicago AD Template
}
\details{
Thank you to Steve Miller's {stevetemplates} library for providing a baseline for this function.
}
\examples{
try(rmarkdown::render("path/to/your.Rmd", kevins.templates::chicagoad()))
}
